pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "air-cli"
version = "0.1.0"
description = "AIR CLI: human-in-the-loop approval, evidence management, forensic command execution"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Steve Anson"},
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
air = "air_cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
